Fix issue that is encountered between Clang and GCC.
Both Clang and GCC order nested operations differently. In the following
scenario:

foo(a(), b())

Clang extracts the nested operations into the following:

t1 = a()
t2 = b()
foo(t1, t2)

However GCC was extracting the operations like so:

t1 = b()
t2 = a()
foo(t1, t2)

This appears to be implementation specific detail, and down to whether
the order should be as a human would interpret it (Clang) or how the
computer (Intel specifically) would interpret it (GCC).